backend/requirements.txt
accelerate==0.33.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
annoy==1.17.3
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blis==0.7.11
boto3==1.34.160
botocore==1.34.160
bpemb==0.3.5
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
coloredlogs==15.0.1
colorlog==6.8.2
confection==0.1.5
conllu==5.0.1
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
Deprecated==1.2.14
distro==1.9.0
Django==4.2.15
django-cors-headers==4.1.0
django-environ==0.10.0
djangorestframework==3.14.0
dnspython==2.2.1
drf-spectacular==0.27.2
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl
environs==9.5.0
exceptiongroup==1.2.2
Faker==24.4.0
fastapi==0.112.0
fastembed==0.3.4
filelock==3.15.4
flair==0.13.1
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
gdown==5.2.0
gensim==4.3.3
greenlet==3.0.3
grpcio==1.58.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
importlib_resources==6.4.0
inflection==0.5.1
iniconfig==2.0.0
InstructorEmbedding==1.0.1
Janome==0.5.0
jellyfish==1.1.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.2.13
langchain-community==0.2.12
langchain-core==0.2.30
langchain-experimental==0.0.64
langchain-openai==0.1.7
langchain-text-splitters==0.2.2
langcodes==3.4.0
langdetect==1.0.9
langsmith==0.1.99
language_data==1.2.0
lark==1.1.9
llama-index==0.9.24
loguru==0.7.2
lxml==5.3.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.2
mdurl==0.1.2
milvus-lite==2.4.9
mmh3==4.1.0
more-itertools==10.4.0
mpld3==0.5.10
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nanoid==2.0.0
nemoguardrails==0.9.1.1
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.2
numpy==1.26.4
onnx==1.16.2
onnxruntime==1.18.1
openai==1.30.1
orjson==3.10.7
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
phonenumbers==8.13.43
pillow==10.4.0
pluggy==1.5.0
pptree==3.1
preshed==3.0.9
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
prompt_toolkit==3.0.47
protobuf==4.24.4
psutil==6.0.0
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pymilvus==2.4.5
pymongo==4.6.1
pyparsing==3.1.2
PySocks==1.7.1
PyStemmer==2.2.0.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch_revgrad==0.2.0
pytz==2022.4
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
rich==13.7.1
rpds-py==0.20.0
s3transfer==0.10.2
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.10.1
segtok==1.5.11
selenium==4.23.1
semver==3.0.2
sentence-transformers==2.2.2
sentencepiece==0.2.0
shellingham==1.5.4
simpleeval==0.9.13
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.32
sqlitedict==2.1.0
sqlparse==0.5.1
srsly==2.4.8
sse-starlette==2.0.0
starlette==0.37.2
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tldextract==5.1.2
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformer-smaller-training-vocab==0.4.0
transformers==4.44.0
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==1.26.19
uvicorn==0.30.6
wasabi==1.1.3
watchdog==4.0.2
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
Wikipedia-API==0.6.0
wordcloud==1.9.2
wrapt==1.16.0
wsproto==1.2.0
yake==0.4.8
yarl==1.9.4
zipp==3.20.0